server/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-celery-results==2.4.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
gunicorn==20.1.0
emanifest==2.0.6
idna==3.4
psycopg2-binary==2.9.5
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
python-dotenv~=0.21.0
celery==5.2.7
redis==4.3.5
drf-spectacular==0.24.2
